requirements.txt
# Core dependencies
streamlit>=1.31.0
face_recognition>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
dlib>=19.24.0
cmake>=3.25.0
requests>=2.31.0

# Optional cloud storage dependencies (uncomment if needed)
# cloudinary>=1.36.0  # For Cloudinary storage backend
# boto3>=1.28.0       # For AWS S3 storage backend
